Changing to a more positive outlook, we change the option 'skip_rcfile' to 'use_rcfile'.
Fix a bug in the option translation code that didn't properly translate dashes to underscores when the option was an invertered boolean (i.e. an option beginning with 'no' or 'no-').

Modify tests so that Module::Build objects are constructed with 'use_rcfile' set to false so that user resource files do not interfere with tests.

Add brief documentation to describe Global commandline options.


git-svn-id: http://svn.perl.org/modules/Module-Build/trunk@5661 50811bd7-b8ce-0310-adc1-d9db26280581