Removed the gen_manpages, gen_html, install_manpages, install_html properties that were a failed attempt to provide user control of generation of documentation in favor of method introduced below.
Introduced two actions: html & manpages that will generate documentation for POD when the respective feature is enabled. The actions will generate the documents even if there is no set place to install to. However, if the actions are invoked as a dependency of another action (e.g. build), the documentation will only be built when there is a default place to install to.


git-svn-id: http://svn.perl.org/modules/Module-Build/trunk@5664 50811bd7-b8ce-0310-adc1-d9db26280581