The command to delete 'Build.bat' did not work on Win9x because the syntax differs between Win9x and WinNT variants. We now use the appropriate syntax based on Win32::GetOSVersion().
The user can invoke the 'realclean' action with either of 'Build realclean' (the batch script) or 'perl Build realclean'. We need to know which one was used, so we customize the use of the 'pl2bat' utility to append a special option, '--batch_bat', to the command line in the batch script. Based on the presence of this option we use an appropriate method of removing 'Build.bat'.


git-svn-id: http://svn.perl.org/modules/Module-Build/trunk@5785 50811bd7-b8ce-0310-adc1-d9db26280581