change from test.pl to basic.t in sample dist in runthrough.t
The test.pl had an bizarre failure on 5.6.2 trying to check for
blib.  Since runthrough.t is really testing dist build capabilities
we don't want a test that can fail and interrupt that process.
Dropped test.pl and restored basic.t so we have a test that always
will pass.



git-svn-id: http://svn.perl.org/modules/Module-Build/trunk@13178 50811bd7-b8ce-0310-adc1-d9db26280581