Replace YAML.pm with YAML::Tiny
We now use YAML::Tiny in place of YAML.pm.  We lose ordered META.yml
files, but YAML::Tiny is more robust and better maintained. Also,
Parse::CPAN::Meta is based on YAML::Tiny, so now we are writing in the
same YAML dialect it expects.  Module::Build::YAML has been updated
with the guts of YAML::Tiny and there is a nifty developer tool to 
suck in the latest guts of YAML::Tiny, stop it from exporting things,
and s/YAML::Tiny/Module::Build::YAML/.



git-svn-id: http://svn.perl.org/modules/Module-Build/trunk@13301 50811bd7-b8ce-0310-adc1-d9db26280581