Revert changes STOP-INSTALL-MANPAGES from the old commit 876e97c48
The old commit 876e97c48... had changes to improve HTML generation on ActivePerl.

Most changes were ok, but not all. In fact one change proved to be problematic
with respect to testing in t/manifypods.t on windows:

Here is the test log:

# Failed test 'Created t\...\man\man1\haspod.pl.1 manpage'     at t/manifypods.t line 134.
# Failed test 'Created t\...\man\man3\Simple.3 manpage'        at t/manifypods.t line 134.
# Failed test 'Created t\...\man\man3\Simple.AllPod.3 manpage' at t/manifypods.t line 134.
#
# Looks like you failed 3 tests of 21.

This commit reverts only this one change, all the other changes of the old commit 876e97c48
remain intact.

That original change (which is now reverted by this commit) was:

+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
++ sub install_map (line 4830)
++
++ Added some logic so that man pages are not installed unless you are
++ running on a Unix like operating system.
+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++