M::B::Compat better handling of dotted-decimal prereqs
I've been using M::B::Compat for a while and saw this happen from time
to time:

Prereq '0.1.2' for 'File::Find::Object' is not supported by
Module::Build::Compat

I meant to patch it so the message is more helpful for us authors who
get perplexed at first. Then, after some digging around the guts of
M::B, I realized it could auto-convert the "common" dotted-decimal
case to the proper format. 0.1.2 -> 0.001002

Here's the patch that adds the auto-conversion logic and some tests
for it. However, I'm not entirely sure the conversion logic is ok, and
one of my test blocks FAILs. ( patch applied against SVN r14210 )

I'm hoping this is a good building block for you to use and tweak
until it's working :)

P.S. Thanks goes to rindolf ( Shlomi Fish ), who reminded me of this
issue and made me submit this :)