Deprecate ModuleInfo and Version modules
We now rely on external dependency Module::Metadata
(which is an extracted copy of Module::Build::Version).

In order to bootstrap, we list that and version.pm as
configure_require dependencies and provide bootstrap copies
in inc/.

M::B::ModuleInfo and M::B::Version are deprecated and now
exist just as simple subclass wrappers for compatibility.