Fix failing test under a UID without passwd entry
As reported in this Debian bug:
<http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=624850> some of the
tests in t/tilde.t fail when run under a UID without a passwd entry.
This can cause the build to fail in some automated environments. The
attached patch against 0.3800 fixes this by testing for definedness in
the results from getpwuid. A similar patch against 0.3603 has been
tested.

Signed-off-by: David Golden <dagolden@cpan.org>