PodParser now respects =encoding directives
When deriving authors or abstract from pod, if encoding is ignored the
related META files will then be incorrect.

This implementation makes reading the pod two pass, first to find an
=encoding directive if it exists and if so set the binmode, then seeks
back to the top to continue parsing as usual.

To support this I have also moved away from testing with a tied class,
which was then mocking out everything that was attempting to be tested.
It now simply uses open to a string.