Preprocess file lists generated by rscan_dir() to sort them
The rscan_dir() function traverses a directory with File::Find,
which returns files in readdir() order. This order is nondeterministic
and depends on the file system.

The lists are used, among other things, to find C files to compile
(in process_support_files()) and later to link (in c_link()).
The linking order affects the generated binary, essentially rendering
it nondeterministic and breaking reproducibility.

Bug-Debian: https://bugs.debian.org/797709
Bug: https://rt.cpan.org/Public/Bug/Display.html?id=106813